refactor: extract SubsonicArtistInfoController — fixes #106
